Removed Offset Argument from Builder Methods
Changed MazeBuilder and various decorators to remove the common offset
argument. It was much easier to add support for a translate() method to
a few classes and move the whole design to its correct location instead
of passing around an offset everywhere. This change should make
everything simpler and cleaner.