Improvements to Door Decorators
* Added a check so that DecoratorHallwayDoor does not place doors in
hallways with ceiling holes. Although unlikely, it's possible that allow
that case could conceal the exit to an area behind a door.
* Changed DecoratorWallDoors so that the two doors are placed with a gap
between them if the room is large enough. It looks better and keeps them
symmetrical regardless of room size.
* Also forgot to mention in the previous commit that I commented out the
code for building the top face of each room. It's very useful for
testing whether decorators are working properly by allowing me to look
into the maze from outside.