Fixed Decay Effects for Mazes
* Fixed the bug with decay effects. It turns out that the partition tree
wasn't being translated correctly. Nodes now translate their children
recursively.
* Tweaked the number of decay operations to be applied. Although that's
not actually important because that value will be configurable in the
future to suit the structure being generated.
* Cut out most of the code in DoorwayData by extending BoundingBox -
almost every function in DoorwayData was also implemented by
BoundingBox.

Known issues: Confirmed that there is a bug in section linking that
causes an exception. It's difficult to track down because it affects
less than 1 in 10,000 designs.